requirements.txt
asgiref==3.8.1
astroid==3.3.8
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cryptography==44.0.0
dill==0.3.9
Django==5.1.3
flake8==7.1.1
idna==3.10
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
pylint==3.3.3
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sqlparse==0.5.2
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
